requirements.txt
Flask
Flask-RESTful
Flask-JWT
psycopg2-binary
requests
